requirements.txt
pandas
openpyxl
python-docx
pydantic>=2
openai
python-dotenv

